lumieresecrete/apps/api/routers.py
from rest_framework import routers
from .views import UserViewSet, ProductViewSet, OrderViewSet

router = routers.DefaultRouter()
router.register(r'users', UserViewSet)
router.register(r'products', ProductViewSet)
router.register(r'orders', OrderViewSet)

urlpatterns = router.urls