lumieresecrete/requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.5
django-cors-headers==3.14.0
django-environ==0.8.1
pandas==1.5.3
openpyxl==3.0.10
matplotlib==3.6.3
django-filter==22.1
django-extensions==3.2.1
pytest==7.2.2
pytest-django==4.5.2
black==22.3.0
flake8==4.0.1
mypy==0.991
docker==5.0.3
djangorestframework-simplejwt==5.2.2
WeasyPrint==59.0
qrcode[pil]==7.4.2
